classification-pu9730962-main/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.5=py37h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2022.10.11=haa95532_0
certifi=2022.9.24=py37haa95532_0
cffi=1.15.1=py37h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.5=py37haa95532_0
cryptography=38.0.1=py37h21b164f_0
cuda=11.6.2=0
cuda-cccl=11.6.55=0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=0
cuda-cudart-dev=11.6.55=0
cuda-cuobjdump=11.6.124=0
cuda-cupti=11.6.124=0
cuda-cuxxfilt=11.6.124=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=0
cuda-nvrtc=11.6.124=0
cuda-nvrtc-dev=11.6.124=0
cuda-nvtx=11.6.124=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py37hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
eigen=3.3.7=h59b6b97_1
entrypoints=0.4=py37haa95532_0
fftw=3.3.9=h2bbff1b_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
glib=2.69.1=h5dc1a3c_1
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.12.1=h1756f20_2
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py37haa95532_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.2=py37haa95532_0
ipython=7.31.1=py37haa95532_1
jedi=0.18.1=py37haa95532_1
joblib=1.1.1=py37haa95532_0
jpeg=9e=h2bbff1b_0
jupyter_client=7.3.5=py37haa95532_0
jupyter_core=4.11.1=py37haa95532_0
kiwisolver=1.4.2=py37hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_4
libiconv=1.16=h2bbff1b_2
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.4.0=h8a3f274_1
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib-base=3.5.3=py37hd77b12b_0
matplotlib-inline=0.1.6=py37haa95532_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
munkres=1.1.4=py_0
nest-asyncio=1.5.5=py37haa95532_0
nsight-compute=2022.3.0.22=0
numexpr=2.8.3=py37hb80d3ca_0
numpy=1.21.5=py37h7a0a035_3
numpy-base=1.21.5=py37hca35cd5_3
opencv=4.6.0=py37h104de81_2
openssl=1.1.1s=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h6214cd6_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py37hdc2b20a_1
pip=22.2.2=py37haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.9.0=py37h2bbff1b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37haa95532_0
pysocks=1.7.1=py37_1
python=3.7.13=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.13.0=py3.7_cuda11.6_cudnn8_0
pytorch-cuda=11.6=h867d48c_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py37haa95532_0
pywin32=302=py37h2bbff1b_2
pyzmq=23.2.0=py37hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py37haa95532_0
scikit-learn=1.0.2=py37hf11a4ad_1
scipy=1.7.3=py37h7a0a035_2
seaborn=0.12.0=py37haa95532_0
setuptools=65.5.0=py37haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h2bbff1b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
torchaudio=0.13.0=pypi_0
torchvision=0.14.0=pypi_0
tornado=6.2=py37h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.3.0=py37haa95532_0
urllib3=1.26.12=py37haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
xz=5.2.6=h8cc25b3_0
zeromq=4.3.4=hd77b12b_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
